Move some tests to a separate file
